+++ src/assets/static/js/wptelegram--main.121b54e4.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('jquery', 'react', 'react-dom', 'wp-api-fetch', 'wp-i18n', 'wp-polyfill'), 'version' => '0b046618db2a76f2f73aaa7d483e53ff');